requirements.txt
alembic==1.4.2
blinker==1.4
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
email-validator==1.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
GoogleAppEngineCloudStorageClient==1.9.22.1
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
Pillow==7.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
Werkzeug==1.0.1
Whoosh==2.7.4
WTForms==2.3.1
